Split up the smokes tests into their own individual files
